Fix for importing a course via :pathname
The API allows importing a course via multipart/form-data. We'd
implemented this, but the tests were insufficient due to the fact SCORM
will accept invalid data immediately, and only return an error if you
check on it a bit later.

The bug we had was that we were still passing `courseName` in the body
as well. This, while required (but ignored) for the `url` option can't
be present for the `pathname` option as SCORM would see that first and
assume it's the zip file contents, which obviously it isn't.

This fixes that.