add endpoints for app configuration in API v2
The 20.1 version of Rustici Engine allows changing application
configuration via the v2 API. This commit adds new methods to the client
for getting and setting configuration.

As a requisite, it also enhances the Request module (and underlying
methods) to support making requests against the v2 API, which not only
has a new path, but also requires the tenant ID as a request header
instead of in the path itself like the v1 API did.

I also took the opportunity to update the version of faraday we're using
to the latest and add myself as an author of the gem.