update specs against 20.1 engine, re-recording vcrs
This updates the gem and specs to handle some slight variations in
response expected when connecting to a 20.1 version of SCORM Engine. The
only functional difference is that this new engine returns URLs back
into itself as relative instead of absolute (as it did before).

In addition to that, many of the error messages are slightly different
in this new version so I've updated the specs testing against them to be
more flexible.